This commit was generated by cvs2svn to compensate for changes in r8275, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@8276 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/monitor/mon_assemble6502.c
@@ -118,7 +118,8 @@ static int mon_assemble_instr(const char *opcode_name, unsigned int operand)
         return -1;
     }
 
-    len = (monitor_cpu_type.asm_addr_mode_get_size)(operand_mode, 0, 0);
+    len = (monitor_cpu_type.asm_addr_mode_get_size)
+          ((unsigned int)(operand_mode), 0, 0);
 
     /* EP 98.08.23 use correct memspace for assembling.  */
     mon_set_mem_val(mem, loc, opcode);
